Rename internal/profile handler tests to pair with their source files
